Fix display of cr2

+++ kernel/src/interrupts.cpp
@@ -64,8 +64,8 @@ void idt_set_gate(size_t gate, void (*function)(void), uint16_t gdt_selector, id
     entry.offset_high= function_address  >> 32;
 }
 
-uint16_t get_cr2(){
-    uint16_t value;
+uint64_t get_cr2(){
+    uint64_t value;
     __asm__ __volatile__("mov rax, cr2; mov %0, rax;" : "=m" (value));
     return value;
 }
